remove tokenInAmount, sweep excess tokenIn from adapter
tokenInAmount doesn't need to be included in the request since the uni v3 router calldata will transfer the tokenIn that it needs. If too much tokenIn is given to the adapter, it needs to be swept at the end of the call to the adapter owner.

modified tests to use the new fn signature and to check for zero balance of all tokens in the adapter